Add Socket Server & Client/Daemon/Irc Message Queues
Add a socket server that simply decodes data from the client and writes
it to a TQueue, while sending data to the client from another TQueue.

The Client Queue contains messages for a connected Client. The Daemon
Queue contains messages for the Daemon from clients. The Irc Queue
contains messages for the Daemon from IRC servers.

After forking the socket server, the main thread will listen to the
Daemon & Irc Queues, responding to any messages they receive.

The IRCState is now saved when connecting to IRC servers, so that the
main thread can send messages to the servers when it is processing the
message queues.