Add Per-Client Queues & Channel Subscriptions
Add a ClientId TVar to the Server, tracking the next Client's ID. When a
new socket connection is made, this value is fetched & then incremented.

Replace the Server's ClientQueue with a Map from ClientIds to
ClientQueues. Messages for Clients are tagged with the ClientId they are
for.

Add a Subscriptions TVar to the Server, mapping from a Server's Channel
to the list of Clients subscribed to that server. Add a `Subscribe`
Daemon message for Clients to subscribe to a list of Channels with.
Broadcast any NewMessage & SendMessage requests to all subscribed Clients.

Use TMQueues for the Client Queue so that Queues can be closed when a
Client has disconnected.

Store the entire ServerMap in a TVar instead of just the ServerData.
This allows us to add Servers to the map as we connect to them.