Store & Render Channel Topics
When a channel's topic is changed, store it in the ChannelData & render
it in the client. Topic changes create messages in the log & are always
shown in a header banner above the message log.

Note that the daemon does not yet parse the topic sent upon joining the
channel, it must be changed in order to be picked up.

* Add the `TopicMessage` constructor to the `ChannelMessage` type, to
  differentiate chat messages from topic change messages.